+++ Changelog.md
@@ -1,6 +1,9 @@
 # Next
 
 - **Breaking Change** Renamed `ReactiveCocoaMoyaProvider` to `ReactiveSwiftMoyaProvider`.
+- **Breaking Change** Renamed `filterStatusCodes(:)` to `filter(statusCodes:)` (and `filterStatusCode(:)` to `filter(statusCode:)`).
+- **Breaking Change** Renamed `request(token:)` to simply `request(:_)` (ReactiveSwift).
+- **Breaking Change** Renamed `notifyPluginsOfImpendingStub(request:)` to `notifyPluginsOfImpendingStub(for:)`.
 - Renamed the `ReactiveCocoa` subspec to `ReactiveSwift`.
 - `PluginType` can now modify requests and responses through `prepareRequest` and `processResponse`